chore(hive): task 37 global sanitization and dna consolidation
- Moved RuleBasedStrategy to Transformer nucleotide (rule_strategy.py).
- Moved MarketService to Connector nucleotide (market.py) with backward compatibility in services/__init__.py.
- Updated aura-core dna.py with correct fractal paths and Toolshed registration.
- Fixed Pattern Heresies (raw print and os.getenv) in docs, notebooks, and telegram-bot.
- Standardized imports in train_dspy.py.
- Verified with full repository tests and bee.Keeper audit (Status: PURE).
