+++ core-service/scripts/training/train_dspy.py
@@ -15,7 +15,7 @@ import structlog
 from dspy.teleprompt import BootstrapFewShot
 
 # Add src to path for imports
-sys.path.append(str(Path(__file__).parent / "src"))
+sys.path.append(str(Path(__file__).parent.parent.parent))
 
 from src.hive.transformer.engine import AuraNegotiator
 from src.hive.transformer.prepare.clean import clean_and_parse_json
