+++ packages/aura-core/src/aura_core/dna.py
@@ -189,10 +189,11 @@ ALLOWED_CHAMBERS = {
     "core-service/migrations": "HiveEvolutionaryScrolls",
     "core-service/tests": "ValidationPollen",
     "api-gateway": "HiveGate",
-    "core-service/src/config": "SacredCodex",
+    "core-service/src/hive/metabolism": "SacredCodex",
+    "core-service/src/hive/metabolism/config": "SacredCodex",
     "core-service/src/services": "WorkerDirectives",
-    "core-service/src/llm": "ReasoningNucleus",
-    "core-service/src/crypto": "SecurityCitadel",
+    "core-service/src/hive/transformer": "ReasoningNucleus",
+    "core-service/src/hive/connector": "SecurityCitadel",
     "core-service/src/prompts": "EchoChamber",
     "core-service/src/guard": "HiveMembrane",
     "deploy": "HiveArmor",
@@ -201,7 +202,7 @@ ALLOWED_CHAMBERS = {
     "agents": "WorkerCells",
     "adapters": "HiveExtensions",
     "frontend": "HiveWindow",
-    "tools": "ToolShed",
+    "tools": "Toolshed",
     "tests": "OuterValidationPollen",
     "packages": "HiveNucleotides",
 }
